move all merge related apis into their own routes file
